@@ -1,3 +1,10 @@
 module github.com/rcarmo/go-te
 
-go 1.22
+go 1.24.0
+
+require (
+	github.com/clipperhouse/uax29/v2 v2.2.0 // indirect
+	github.com/mattn/go-runewidth v0.0.19 // indirect
+	github.com/rivo/uniseg v0.4.7 // indirect
+	golang.org/x/text v0.34.0 // indirect
+)
