@@ -6,30 +6,35 @@ const (
 	ColorDefault ColorMode = iota
 	ColorANSI16
 	ColorANSI256
+	ColorTrueColor
 )
 
 type Color struct {
 	Mode  ColorMode
 	Index uint8
+	Name  string
 }
 
 type Attr struct {
-	Fg        Color
-	Bg        Color
-	Bold      bool
-	Underline bool
-	Blink     bool
-	Reverse   bool
-	Conceal   bool
+	Fg            Color
+	Bg            Color
+	Bold          bool
+	Italics       bool
+	Underline     bool
+	Strikethrough bool
+	Reverse       bool
+	Blink         bool
+	Conceal       bool
 }
 
 type Cell struct {
-	Ch   rune
+	Data string
 	Attr Attr
 }
 
 type Cursor struct {
 	Row    int
 	Col    int
+	Attr   Attr
 	Hidden bool
 }
